requirements.txt
Flask==0.12
Flask-PyMongo==0.4.1
Flask-Script==2.0.5
gunicorn==19.7.0
celery==4.0.2
sh==1.12.9
Flask-RESTful==0.3.5
redis==2.10.5
requests==2.18.1
PyMySQL==0.7.11
Flask-Migrate==2.0.4
Flask-SQLAlchemy==2.2
Flask-Login==0.4.0
Flask-Cache==0.13.1